Move server code into separate file so can test without running server
