+++ package.json
@@ -25,7 +25,6 @@
     "bluebird": "^3.5.2",
     "body-parser": "^1.18.3",
     "express": "^4.16.4",
-    "faker": "^4.1.0",
     "morgan": "^1.9.1",
     "mysql2": "^1.6.1",
     "sequelize": "^4.41.0"
@@ -34,11 +33,15 @@
     "@babel/core": "^7.1.2",
     "@babel/preset-env": "^7.1.0",
     "@babel/preset-react": "^7.0.0",
+    "babel-core": "7.0.0-bridge.0",
     "babel-loader": "^8.0.4",
+    "faker": "^4.1.0",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
+    "jest": "^23.6.0",
     "react": "^16.6.0",
     "react-dom": "^16.6.0",
+    "supertest": "^3.3.0",
     "webpack": "^4.23.1",
     "webpack-cli": "^3.1.2"
   }
